+++ src/v1/HttpStatusCode/mustBeHttpStatusCode.spec.ts
@@ -35,7 +35,7 @@ import { AnyAppError, OnError } from "@ganbarodigital/ts-lib-error-reporting/lib
 import { expect } from "chai";
 import { describe } from "mocha";
 
-import { mustBeHttpStatusCode } from "../internal";
+import { mustBeHttpStatusCode } from ".";
 
 describe("mustBeHttpStatusCode()", () => {
     const onError: OnError = (e: AnyAppError): never => {