+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### Refactor
+
+* Now republishes internal types from `@ganbarodigital/ts-lib-error-reporting`
+  - this was done to address problems with circular dependencies
+
 ## v0.1.3
 
 ### Fixes